// src/common/enums/user-role.enum.ts
export enum UserRole {
  SUPER_ADMIN = 'SUPER_ADMIN',
  FIELD_AGENT = 'FIELD_AGENT',
  PET_OWNER = 'PET_OWNER',
  CLINIC_OWNER = 'CLINIC_OWNER',
  CLINIC_MANAGER = 'CLINIC_MANAGER',
  RECEPTIONIST = 'RECEPTIONIST',
  ASSISTANT = 'ASSISTANT',
  SHOP_OWNER = 'SHOP_OWNER',
  DELIVERY_PARTNER = 'DELIVERY_PARTNER',
}